调整上传数据位数

+++ src/models.rs
@@ -22,8 +22,8 @@ pub struct Target {
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PingData {
     pub ipv6: bool,
-    pub min: f64,
-    pub jitter: f64,
+    pub min: f32,
+    pub jitter: f32,
     pub failed: i32,
 }
